# Speed up the state tests somewhat.