--- a/state/open.go
+++ b/state/open.go
@@ -10,6 +10,7 @@ import (
 
 	"github.com/juju/errors"
 	"github.com/juju/utils"
+	"github.com/juju/utils/clock"
 	"gopkg.in/juju/names.v2"
 	"gopkg.in/mgo.v2"
 	"gopkg.in/mgo.v2/txn"
@@ -41,7 +42,7 @@ func Open(
 	info *mongo.MongoInfo, opts mongo.DialOpts,
 	newPolicy NewPolicyFunc,
 ) (*State, error) {
-	st, err := open(controllerModelTag, info, opts, newPolicy)
+	st, err := open(controllerModelTag, info, opts, newPolicy, clock.WallClock)
 	if err != nil {
 		return nil, errors.Trace(err)
 	}
@@ -64,6 +65,7 @@ func open(
 	controllerModelTag names.ModelTag,
 	info *mongo.MongoInfo, opts mongo.DialOpts,
 	newPolicy NewPolicyFunc,
+	clock clock.Clock,
 ) (*State, error) {
 	logger.Infof("opening state, mongo addresses: %q; entity %v", info.Addrs, info.Tag)
 	logger.Debugf("dialing mongo")
@@ -80,7 +82,7 @@ func open(
 	}
 	logger.Debugf("mongodb login successful")
 
-	st, err := newState(controllerModelTag, controllerModelTag, session, info, newPolicy)
+	st, err := newState(controllerModelTag, controllerModelTag, session, info, newPolicy, clock)
 	if err != nil {
 		return nil, errors.Trace(err)
 	}
@@ -104,6 +106,10 @@ func mongodbLogin(session *mgo.Session, mongoInfo *mongo.MongoInfo) error {
 
 // InitializeParams contains the parameters for initializing the state database.
 type InitializeParams struct {
+	// Clock wraps all calls time. Real uses use clock.WallClock,
+	// tests may override with a testing clock.
+	Clock clock.Clock
+
 	// ControllerModelArgs contains the arguments for creating
 	// the controller model.
 	ControllerModelArgs ModelArgs
@@ -147,6 +153,9 @@ type InitializeParams struct {
 
 // Validate checks that the state initialization parameters are valid.
 func (p InitializeParams) Validate() error {
+	if p.Clock == nil {
+		return errors.NotValidf("missing clock")
+	}
 	if err := p.ControllerModelArgs.Validate(); err != nil {
 		return errors.Trace(err)
 	}
@@ -202,7 +211,7 @@ func Initialize(args InitializeParams) (_ *State, err error) {
 	// When creating the controller model, the new model
 	// UUID is also used as the controller UUID.
 	modelTag := names.NewModelTag(args.ControllerModelArgs.Config.UUID())
-	st, err := open(modelTag, args.MongoInfo, args.MongoDialOpts, args.NewPolicy)
+	st, err := open(modelTag, args.MongoInfo, args.MongoDialOpts, args.NewPolicy, args.Clock)
 	if err != nil {
 		return nil, errors.Trace(err)
 	}
@@ -458,6 +467,7 @@ func newState(
 	modelTag, controllerModelTag names.ModelTag,
 	session *mgo.Session, mongoInfo *mongo.MongoInfo,
 	newPolicy NewPolicyFunc,
+	clock clock.Clock,
 ) (_ *State, err error) {
 
 	defer func() {
@@ -478,6 +488,7 @@ func newState(
 
 	// Create State.
 	st := &State{
+		clock:              clock,
 		modelTag:           modelTag,
 		controllerModelTag: controllerModelTag,
 		mongoInfo:          mongoInfo,