--- a/state/raceoff_test.go
+++ b/state/raceoff_test.go
@@ -1,0 +1,8 @@
+// Copyright 2012, 2013 Canonical Ltd.
+// Licensed under the AGPLv3, see LICENCE file for details.
+
+// +build !race
+
+package state_test
+
+const raceDetector = false