--- a/state/raceon_test.go
+++ b/state/raceon_test.go
@@ -1,0 +1,8 @@
+// Copyright 2012, 2013 Canonical Ltd.
+// Licensed under the AGPLv3, see LICENCE file for details.
+
+// +build race
+
+package state_test
+
+const raceDetector = true