--- a/state/volume.go
+++ b/state/volume.go
@@ -747,9 +747,8 @@ func (st *State) addVolumeOps(params VolumeParams, machineId string) ([]txn.Op, 
 		return nil, names.VolumeTag{}, errors.Annotate(err, "cannot generate volume name")
 	}
 	status := statusDoc{
-		Status: status.Pending,
-		// TODO(fwereade): 2016-03-17 lp:1558657
-		Updated: time.Now().UnixNano(),
+		Status:  status.Pending,
+		Updated: st.clock.Now().UnixNano(),
 	}
 	doc := volumeDoc{
 		Name:      name,